Cleaned-Up view.py
