--- backend/api/Testing.py
@@ -1,8 +1,0 @@
-def plustwo(n):
-    out = n + 2
-    return out
-
-
-def falldist(t,g=9.81):
-    d = 0.5 * g * t**2
-    return d
