+++ backend/api/urls.py
@@ -6,5 +6,4 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', views.main, name = "Kiosk"),
     path('training/', views.training, name = "Training Model"),
-    path('test/', views.test, name = "Test")
 ]
